backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.1
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
gunicorn>=21.2.0
